backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
supabase==2.3.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI Integration
anthropic==0.18.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# CORS
python-cors==1.0.0
